Completed the guide, need real changes now

+++ lib/models/baby.dart
@@ -1,10 +1,9 @@
 class Baby {
-  
   final String name;
   final int age;
   final String gender;
-  final int weight;
-  final int height;
+  final double weight;
+  final double height;
 
   Baby({
     required this.name,
@@ -13,5 +12,4 @@ class Baby {
     required this.weight,
     required this.height,
   });
-
 }
